Admin

+++ desktop/src/renderer/main.js
@@ -1,6 +1,9 @@
 import Vue from 'vue'
 import axios from 'axios'
 import Buefy from 'buefy'
+import VueMoment from 'vue-moment'
+import VueCookie from 'vue-cookie'
+import moment from 'moment-timezone'
 
 import App from './App'
 import router from './router'
@@ -9,32 +12,32 @@ import store from './store'
 import state from './global'
 Vue.prototype.$global = state
 
-const baseUrl = {
+const url = {
   development: 'http://localhost:3333',
   production: 'https://trellu-app.herokuapp.com'
 }
 
 if (!process.env.IS_WEB) Vue.use(require('vue-electron'))
-const api = axios.create({ baseURL: baseUrl[process.env.NODE_ENV] })
+const api = axios.create({ baseURL: url[process.env.NODE_ENV] })
 api.interceptors.response.use(
   response => response,
   error => {
-    switch (error.response.data.error.name) {
-      case 'InvalidSessionException':
-        localStorage.clear()
-        router.replace({path: '/'})
-        break
-      case 'HttpException':
-        // api.get('api/auth/logout')
-        break
-      default:
-        break
+    // console.log(error.response.data.error.message)
+    if (error.response.status === 401) {
+      localStorage.clear()
+      router.replace({path: '/'})
+    } else {
+      return Promise.reject(error.response.data)
     }
   }
 )
 
 Vue.api = Vue.prototype.$api = api
 Vue.config.productionTip = false
+Vue.use(VueMoment, {
+  moment
+})
+Vue.use(VueCookie)
 Vue.use(Buefy, {
   defaultIconPack: 'mdi'
 })
